feat: Implement editor functionality with resizable panels and project management
- Added Editor component with resizable panels for media, properties, preview, and timeline.
- Integrated project loading and creation logic based on project ID from URL.
- Created layout component for locale handling and static rendering.
- Developed home page with dynamic metadata and waitlist count fetching.
- Implemented projects page with project creation, selection, and bulk deletion features.
- Added Why Not Capcut page with satirical content on video editing apps.
- Introduced language switcher component for locale selection.
- Established navigation and request handling for internationalization.
- Defined routing and types for next-intl integration.
