+++ apps/web/package.json
@@ -37,6 +37,7 @@
     "lucide-react": "^0.468.0",
     "motion": "^12.18.1",
     "next": "^15.3.4",
+    "next-intl": "^4.3.4",
     "next-themes": "^0.4.4",
     "pg": "^8.16.2",
     "radix-ui": "^1.4.2",
