+++ apps/web/src/app/[locale]/api/auth/[...all]/route.ts
@@ -1,0 +1,4 @@
+import { auth } from "@opencut/auth/server";
+import { toNextJsHandler } from "better-auth/next-js";
+
+export const { POST, GET } = toNextJsHandler(auth);
