+++ apps/web/src/app/[locale]/api/health/route.ts
