--- apps/web/src/app/api/auth/[...all]/route.ts
@@ -1,4 +1,0 @@
-import { auth } from "@opencut/auth/server";
-import { toNextJsHandler } from "better-auth/next-js";
-
-export const { POST, GET } = toNextJsHandler(auth);
