+++ apps/web/src/middleware.ts
@@ -1,5 +1,10 @@
 import { NextResponse } from "next/server";
 import type { NextRequest } from "next/server";
+import createMiddleware from 'next-intl/middleware';
+import { routing } from './i18n/routing';
+
+// 创建 next-intl 中间件
+const handleI18nRouting = createMiddleware(routing);
 
 export async function middleware(request: NextRequest) {
   // Handle fuckcapcut.com domain redirect
@@ -9,24 +14,48 @@ export async function middleware(request: NextRequest) {
 
   const path = request.nextUrl.pathname;
 
-  if (path === "/editor" && process.env.NODE_ENV === "production") {
-    const homeUrl = new URL("/", request.url);
-    homeUrl.searchParams.set("redirect", request.url);
-    return NextResponse.redirect(homeUrl);
+  // 跳过 next-intl 处理的特定路径
+  if (
+    path.startsWith('/api') ||
+    path.startsWith('/_next') ||
+    path.startsWith('/_vercel') ||
+    path.includes('.')
+  ) {
+    if (path === "/editor" && process.env.NODE_ENV === "production") {
+      const homeUrl = new URL("/", request.url);
+      homeUrl.searchParams.set("redirect", request.url);
+      return NextResponse.redirect(homeUrl);
+    }
+    return NextResponse.next();
   }
 
-  return NextResponse.next();
+  // 应用 next-intl 路由处理
+  const response = handleI18nRouting(request);
+
+  // 对编辑器路径进行额外检查（在本地化之后）
+  if (response.status === 200) {
+    const url = new URL(request.url);
+    const pathname = url.pathname;
+    
+    // 检查是否是编辑器路径（可能有语言前缀）
+    if (pathname.endsWith('/editor') && process.env.NODE_ENV === "production") {
+      const homeUrl = new URL("/", request.url);
+      homeUrl.searchParams.set("redirect", request.url);
+      return NextResponse.redirect(homeUrl);
+    }
+  }
+
+  return response;
 }
 
 export const config = {
+  // 匹配所有路径名，除了以下开头的：
+  // - api (API 路由)
+  // - _next/static (静态文件)
+  // - _next/image (图像优化文件)
+  // - _vercel (Vercel 内部文件)
+  // - favicon.ico 等包含点的文件
   matcher: [
-    /*
-     * Match all request paths except for the ones starting with:
-     * - api (API routes)
-     * - _next/static (static files)
-     * - _next/image (image optimization files)
-     * - favicon.ico (favicon file)
-     */
-    "/((?!api|_next/static|_next/image|favicon.ico).*)",
+    '/((?!api|_next/static|_next/image|_vercel|.*\\..*).*)'
   ],
 };
